modules/sound-detector/pyproject.toml
[tool.poetry]
name = "sound-detector"
version = "0.1.0"
description = ""
authors = ["Ramon Blanquer <blanquer.ramon@gmail.com>"]

[tool.poetry.dependencies]
pandas = "^2.1.3"
python = "^3.10,<3.11"
pyaudio = "^0.2.14"
pyzmq = "^25.1.1"
tflite-runtime = { version = "~2.13.0", platform = "linux" }

# You can install all the tensorflow (not only the tensorflow lite) by passing `--with tensorflow-dev`
# when running `poetry install`.
python-slugify = "^8.0.1"
influxdb-client = "^1.39.0"
environs = "^10.0.0"
[tool.poetry.group.tensorflow-dev]
optional = true

[tool.poetry.group.tensorflow-dev.dependencies]
# Issue between poetry and tensorflow metadata since >=2.11. This is a temporary
# workaround. Inspired from https://github.com/tensorflow/tensorflow/blob/adb39b04e9cb116df4659a7e2de9eea27e62f25c/tensorflow/tools/pip_package/setup.py#L148-L162
#
#   https://github.com/python-poetry/poetry/issues/8271
#
tensorflow = {version = "~2.13.0" }
tensorflow-macos = { version = "~2.13.0", platform = "darwin", markers = "platform_machine=='arm64'" }
tensorflow-intel = { version = "~2.13.0", platform = "win32" }
tensorflow-cpu = [
    { version = "~2.13.0", platform = "linux", markers = "platform_machine!='arm64' and platform_machine!='aarch64'" },
    { version = "~2.13.0", platform = "darwin", markers = "platform_machine!='arm64' and platform_machine!='aarch64'" }
]
tensorflow-cpu-aws = { version = "~2.13.0", platform = "linux", markers = "platform_machine=='arm64' or platform_machine=='aarch64'" }
tensorflow-io = "~0.34.0"
tensorflow-hub = "0.16.1"

[tool.poetry.group.dev.dependencies]
matplotlib = "^3.8.1"
black = "^23.11.0"
ruff = "^0.1.5"
ipython = "^8.17.2"
notebook = "^7.0.6"
pynvim = "^0.4.3"
pyright = "^1.1.337"
jupyterlab = "^4.0.10"
pytest = "^7.4.4"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
